+++ packages/inertia/example/inertia_client_app/lib/app.dart
@@ -1,7 +1,7 @@
 import 'dart:convert';
 import 'dart:io';
 
-import 'package:inertia_dart/inertia.dart';
+import 'package:inertia_dart/inertia_dart.dart';
 
 const String _clientEntry = 'src/main.jsx';