+++ packages/inertia/example/inertia_ssr_app/bin/ssr.dart
@@ -1,6 +1,6 @@
 import 'dart:io';
 
-import 'package:inertia_dart/inertia.dart';
+import 'package:inertia_dart/inertia_dart.dart';
 
 Future<void> main(List<String> args) async {
   final runtime = Platform.environment['INERTIA_SSR_RUNTIME'] ?? 'node';