+++ packages/inertia/test/unit/inertia_http_test.dart
@@ -4,7 +4,7 @@ library;
 import 'dart:convert';
 import 'dart:io';
 
-import 'package:inertia_dart/inertia.dart';
+import 'package:inertia_dart/inertia_dart.dart';
 import 'package:test/test.dart';
 
 void main() {