+++ packages/inertia/test/unit/middleware_test.dart
@@ -2,7 +2,7 @@
 library;
 
 import 'package:test/test.dart';
-import 'package:inertia_dart/inertia.dart';
+import 'package:inertia_dart/inertia_dart.dart';
 
 /// Runs middleware unit tests.
 void main() {