+++ packages/inertia/test/unit/ssr_server_test.dart
@@ -5,7 +5,7 @@ import 'dart:io';
 
 import 'package:http/http.dart' as http;
 import 'package:http/testing.dart';
-import 'package:inertia_dart/inertia.dart';
+import 'package:inertia_dart/inertia_dart.dart';
 import 'package:path/path.dart' as path;
 import 'package:test/test.dart';