+++ packages/inertia/test/unit/testing_extensions_test.dart
@@ -5,7 +5,7 @@ import 'dart:convert';
 import 'dart:io';
 
 import 'package:server_testing/server_testing.dart';
-import 'package:inertia_dart/inertia.dart';
+import 'package:inertia_dart/inertia_dart.dart';
 
 /// Runs testing helper unit tests.
 void main() {